功能实现

+++ src/main/java/org/wekit/web/db/dao/impl/CodeApplyLogDaoImpl.java
@@ -23,7 +23,7 @@ public class CodeApplyLogDaoImpl extends HibernateBaseDao<CodeApplyLog, Long> im
 	}
 
 	@Override
-	public CodeApplyLog saveCodeApplyLog(String userid, String username, String deptId, String deptname, int fileType, String code, String content, String operateType, long createTime) {
+	public CodeApplyLog saveCodeApplyLog(String userid, String username, String deptId, String deptname, long fileType, String code, String content, String operateType, long createTime) {
 		CodeApplyLog codeApplyLog=new CodeApplyLog(userid,username,deptId,deptname,createTime,fileType,code,content,operateType);
 		return codeApplyLog;
 	}
